\name{plot.select}
\alias{plot.select}
\title{
Plot function for S3 class "select"
}
\description{
Plot the optimal graph by model selection. 
}
\usage{
\method{plot}{select}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"select"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{
Xingguo Li, Tuo Zhao, and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{camel.tiger}} and \code{\link{camel.tiger.select}}
}
