\name{plot.sim}
\alias{plot.sim}
\title{
Plot function for S3 class "sim"
}
\description{
Visualize the covariance matrix, the empirical covariance matrix, the adjacency matrix and the graph pattern of the true graph structure.
}
\usage{
\method{plot}{sim}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"sim"} 
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{
Xingguo Li, Tuo Zhao and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{camel.tiger}} and \code{\link{camel.tiger.generator}}
}
