\name{plot.tiger}
\alias{plot.tiger}
\title{
Plot function for S3 class "camel.tiger"
}
\description{
Plot sparsity level information and 3 typical sparse graphs from the graph path.
}
\usage{
\method{plot}{tiger}(x, align = FALSE, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"tiger"} 
}
  \item{align}{
If \code{align = FALSE}, 3 plotted graphs are aligned
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{
Xingguo Li, Tuo Zhao, and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{camel.tiger}} and \code{\link{camel-package}}
}
