\name{print.cmr}
\alias{print.cmr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print a camel.cmr Object
}
\description{
Print a summary of the information about the cmr object.
}
\usage{
\method{print}{cmr}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  The cmr object.
}
  \item{\dots}{
  Additional print options.
}
}
\details{
This call simply outlines the options used for computing a cmr object.
}
%\value{
%}
%\references{
%}

\author{
Xingguo Li, Tuo Zhao, and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
%\note{
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{camel.cmr}} and \code{\link{camel-package}}.
}
