\name{print.sim}
\alias{print.sim}
\title{
Print function for S3 class "sim"
}
\description{
Print the information about the sample size, the dimension, the pattern and sparsity of the true graph structure.
}
\usage{
\method{print}{sim}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class \code{"sim"}
}
  \item{\dots}{
System reserved (No specific usage)
}
}
\author{
Xingguo Li, Tuo Zhao, and Han Liu \cr
Maintainer: Xingguo Li <xingguo.leo@gmail.com>
}
\seealso{
\code{\link{camel.tiger}} and \code{\link{camel.tiger.generator}}
}
